kcov: add support to kernel automotive

+++ kcov/prepare/runtest.sh
@@ -49,14 +49,22 @@ setup()
     log "submit config"
     rhts-submit-log -l $KCOV_CONF
 
-    rpm -q kernel-gcov
+    rpm -q ${KERNEL_GCOV}
     if [ $? -ne 0 ]; then
         # Expects cki repo with kernel rpms to be configured
         log "install gcov data files"
-        dnf install -y kernel-gcov
-        if [ $? -ne 0 ]; then
-            cki_abort_recipe "Cannot install the gcov data files package.";
-            exit
+        if [[ -e /run/ostree-booted ]]; then
+            rpm-ostree install -A ${KERNEL_GCOV}
+            if [ $? -ne 0 ]; then
+                cki_abort_recipe "Cannot install the gcov data files package.";
+                exit
+            fi
+        else
+            dnf install -y ${KERNEL_GCOV}
+            if [ $? -ne 0 ]; then
+                cki_abort_recipe "Cannot install the gcov data files package.";
+                exit
+            fi
         fi
     fi
 
@@ -95,7 +103,7 @@ verify()
 
     ls $GCOV_BASEDIR/
     if [ $? -ne 0 ]; then
-        cki_abort_recipe "kernel-gcov files were not available after reboot"
+        cki_abort_recipe "${KERNEL_GCOV} files were not available after reboot"
         exit
     fi
     pass
